LibRaw.cmake: Register rtengine depends on LibRaw
Properly register the dependency of rtengine on LibRaw,
when building our local copy of LibRaw.  We do this
from inside LibRaw.cmake, which has the build logic for local libraw.

Without this dependency, the build tool would see libraw as an
independent default (or "all") target and will dispatch it in parallel,
and the build success then depends on the outcome of the race
between building libraw in parallel with building the rest of rtengine
(procparams.cc is large and slow to build).

If however only libraw is to be rebuilt, and we only need to relink
rtengine without building slow-to-build parts of it, the build will
break because the libraw build loses the race and the build tool
(ninja/make/...) would attempt to link against libraw before its
build has completed successfully, and would attempt to link if libraw
broke to build.  Reproducer on Unix, in the build directory,
for me on Fedora 42 on a Ryzen 7 1700 (8 cores, 2 threads per core):

cmake --build . && cmake --build . -t clean-libraw && cmake --build -j20